+++ crates/fj-window/src/window.rs
@@ -1,10 +1,7 @@
 //! CAD viewer utility windowing abstraction
 
 use fj_viewer::{Screen, ScreenSize};
-use winit::{
-    event_loop::EventLoop,
-    window::{Fullscreen, WindowBuilder},
-};
+use winit::{event_loop::EventLoop, window::WindowBuilder};
 
 /// Window abstraction providing details such as the width or height and easing initialization.
 pub struct Window(winit::window::Window);
@@ -15,7 +12,6 @@ impl Window {
         let window = WindowBuilder::new()
             .with_title("Fornjot")
             .with_maximized(true)
-            .with_fullscreen(Some(Fullscreen::Borderless(None)))
             .with_decorations(true)
             .with_transparent(false)
             .build(event_loop)?;
